+++ ModMyFactory.Win32/PercievedFileType.cs
@@ -1,0 +1,25 @@
+//  Copyright (C) 2020-2021 Mathis Rech
+//
+//  This program is free software: you can redistribute it and/or modify
+//  it under the terms of the GNU General Public License as published by
+//  the Free Software Foundation, either version 3 of the License, or
+//  (at your option) any later version.
+
+namespace ModMyFactory.Win32
+{
+    public enum PercievedFileType
+    {
+        None,
+        Folder,
+        Text,
+        Image,
+        Audio,
+        Video,
+        Compressed,
+        Document,
+        System,
+        Application,
+        Gamemedia,
+        Contacts,
+    }
+}