+++ Changelog.md
@@ -1,5 +1,7 @@
-### Beta 4.2.7
+### Beta 4.3
 
+- Add UI option to import modpacks
+- Add associations for FMP and FMPA files on Windows
 - Fix Factorio not getting loaded on Mac
 - Fix multiple instances being allowed to run on Mac